venv-requirements.txt
absl-py==0.15.0
alabaster==0.7.13
alembic==1.10.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.1
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.0.0
boto3==1.26.118
botocore==1.29.118
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chainer==7.8.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.4
cmaes==0.9.1
colorama==0.4.6
colorlog==6.7.0
comm==0.1.3
contourpy==1.0.7
cupy-cuda12x==12.0.0
curio==1.6
cycler==0.10.0
Cython==0.29.34
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docrepr==0.2.0
docutils==0.18.1
easydict==1.10
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
faiss-cpu==1.7.4
fastjsonschema==2.16.3
fastreid==1.4.0
fastrlock==0.8.1
filelock==3.10.7
fire==0.5.0
flatbuffers==1.12
fonttools==4.39.3
fqdn==1.5.1
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.54.0
h5py==3.1.0
huggingface-hub==0.13.4
idna==2.10
imageio==2.27.0
imagesize==1.4.1
imgaug==0.4.0
importlib-metadata==6.5.0
importlib-resources==5.12.0
imutils==0.5.4
iniconfig==2.0.0
ipykernel==6.22.0
ipyparallel==8.6.1
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.3.0
mypy-extensions==1.0.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
optuna==3.1.1
outcome==1.2.0
packaging==23.0
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pathspec==0.11.1
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.5
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.6
pycocotools==2.0.6
pycparser==2.21
Pygments==2.15.1
pyparsing==2.4.7
pyrsistent==0.19.3
pytest==6.2.5
pytest-asyncio==0.20.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roboflow==1.0.5
rsa==4.9
s3transfer==0.6.0
sahi==0.11.13
scikit-image==0.17.2
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
Send2Trash==1.8.0
shapely==2.0.1
six==1.15.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4.1
Sphinx==6.1.3
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.9
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.5.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
terminaltables==3.1.10
testpath==0.6.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.3.21
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
torchviz==0.0.2
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
typing-extensions==3.7.4.3
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
wget==3.2
widgetsnbextension==4.0.7
wrapt==1.12.1
yacs==0.1.8
yolov5==7.0.11
zipp==3.15.0
